asst03/requirements.txt
aniso8601==3.0.2
certifi==2018.8.24
chardet==3.0.4
click==6.7
cycler==0.10.0
Flask==1.0.2
Flask-Cors==3.0.6
flask-restplus==0.11.0
gmplot==1.2.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.0
numpy==1.15.2
oauthlib==2.1.0
panda==0.3.1
pandas==0.23.4
Pillow==5.3.0
pymongo==3.7.1
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
scikit-learn==0.20.0
scipy==1.1.0
simplejson==3.16.0
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
